Escape CRLF in message so output isn't corrupt
Depening on the input, the `JsonException` message can contain CR and
LF. These will mess up the output to standard out, so they are now
replaced with \r \n. This does mean that the message may no longer
accurately reflect the contents of the JSON, but that's better than
having corrupt output.

I thought about using a more structured output format, like, say, JSON,
but rejected that as overly heavy for a tool like this. Perhaps in the
future.