Add "Unixy" integration tests
The same test set is used as is used in the Windows integration tests.

The test-list.tsv file was changed to test-list.csv, as sh's `read`
collapses runs of IFS values _if they are whitespace_. Some of the tests
cases don't have any arguments, so this was causing problems. Switching
to CSV avoids this problem, and none of the arguments need commas in
them.