Add website generation from README via Pandoc
The GitHub workflow currently uses Ubuntu 18.04, which only has Pandoc
1.9. The workflow fetches Pandoc 2.9 and extracts it, as the website
build uses some Pandoc 2+ features.

A custom template is used, as the default template uses some script from
CloudFlare, and I want to minimize external dependencies.

Two custom language syntaxes are defined so that elisp and jsonc
highlight in both the Pandoc output and when README.md is rendered on
GitHub. GitHub has built-in support for elisp and jsonc already.